+++ src/table/handlers/TableHandler.ts
@@ -167,8 +167,7 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
 
     if (
       !options.tableEntityProperties ||
-      !options.tableEntityProperties.PartitionKey ||
-      !options.tableEntityProperties.RowKey
+      !options.tableEntityProperties.PartitionKey
     ) {
       throw StorageErrorFactory.getPropertiesNeedValue(context);
     }
@@ -403,8 +402,9 @@ export default class TableHandler extends BaseHandler implements ITableHandler {
     const accountName = tableContext.account;
     const partitionKey = tableContext.partitionKey!; // Get partitionKey from context
     const rowKey = tableContext.rowKey!; // Get rowKey from context
-
+    console.log("asdasdasdad");
     if (!partitionKey || !rowKey) {
+      debugger;
       throw StorageErrorFactory.getPropertiesNeedValue(context);
     }
     if (ifMatch === "" || ifMatch === undefined) {